Remove global jest mocks
