+++ .github/workflows/CI.yml
@@ -8,7 +8,7 @@ env:
   python-version: '3.11'
   java-version: '11'
   gradle-version: '8.0.2'
-  node-version: '20.x'
+  node-version: '18.x'
 
 jobs:
   style-check:
