+++ deployment/cdk/opensearch-service-migration/package.json
@@ -36,7 +36,7 @@
   "scripts": {
     "build": "tsc",
     "watch": "tsc -w",
-    "test": "jest --coverage",
+    "test": "jest  --coverage",
     "clean": "rm -rf dist",
     "release": "npm run build",
     "cdk": "cdk",
@@ -44,14 +44,14 @@
   },
   "devDependencies": {
     "@aws-cdk/aws-servicecatalogappregistry-alpha": "2.117.0-alpha.0",
-    "@types/jest": "^29.2.5",
+    "@types/jest": "^29.2.4",
     "@types/node": "18.11.18",
     "@types/node-forge": "^1.3.1",
     "aws-cdk": "2.117.0",
     "aws-cdk-lib": "2.117.0",
     "constructs": "^10.0.0",
-    "jest": "^29.3.1",
-    "ts-jest": "^29.0.3",
+    "jest": "^29.2.4",
+    "ts-jest": "^29.2.4",
     "ts-node": "^10.9.1",
     "typescript": "~4.9.4"
   },
