+++ deployment/cdk/opensearch-service-migration/test/migration-services-yaml.test.ts
@@ -4,22 +4,30 @@ import { Template, Capture, Match } from "aws-cdk-lib/assertions";
 import { MigrationConsoleStack } from "../lib/service-stacks/migration-console-stack";
 import { createStackComposer } from "./test-utils";
 import * as yaml from 'yaml';
+import {describe, afterEach, beforeEach, test, expect, jest} from '@jest/globals';
 
+jest.mock('aws-cdk-lib/aws-ecr-assets');
+describe('Migration Services YAML Tests', () => {
+  beforeEach(() => {
+    jest.spyOn(ContainerImage, 'fromDockerImageAsset').mockImplementation(() => ContainerImage.fromRegistry("ServiceImage"));
+});
 
-// Mock using local Dockerfile (which may not exist and would fail synthesis)
-jest.mock("aws-cdk-lib/aws-ecr-assets")
-jest.spyOn(ContainerImage, 'fromDockerImageAsset').mockImplementation(() => ContainerImage.fromRegistry("ServiceImage"));
+  afterEach(() => {
+    jest.clearAllMocks();
+    jest.resetModules();
+    jest.restoreAllMocks();
+    jest.resetAllMocks();
+  });
 
-
-test('Test default servicesYaml can be stringified', () => {
+  test('Test default servicesYaml can be stringified', () => {
     const servicesYaml = new ServicesYaml();
     expect(servicesYaml.metrics_source).toBeDefined();
     expect(Object.keys(servicesYaml.metrics_source)).toContain("cloudwatch");
     const yaml = servicesYaml.stringify();
     expect(yaml).toBe("metrics_source:\n  cloudwatch:\n");
-})
+  });
 
-test('Test servicesYaml with target cluster can be stringified', () => {
+  test('Test servicesYaml with target cluster can be stringified', () => {
     let servicesYaml = new ServicesYaml();
     const cluster: ClusterYaml = { 'endpoint': 'https://abc.com', 'no_auth': '' };
     servicesYaml.target_cluster = cluster;
@@ -27,9 +35,9 @@ test('Test servicesYaml with target cluster can be stringified', () => {
     expect(servicesYaml.target_cluster).toBeDefined();
     const yaml = servicesYaml.stringify();
     expect(yaml).toBe(`target_cluster:\n  endpoint: ${cluster.endpoint}\n  no_auth: ""\nmetrics_source:\n  cloudwatch:\n`);
-})
+  });
 
-test('Test servicesYaml with source and target cluster can be stringified', () => {
+  test('Test servicesYaml with source and target cluster can be stringified', () => {
     let servicesYaml = new ServicesYaml();
     const targetCluster: ClusterYaml = { 'endpoint': 'https://abc.com', 'no_auth': '' };
     servicesYaml.target_cluster = targetCluster;
@@ -43,9 +51,9 @@ test('Test servicesYaml with source and target cluster can be stringified', () =
     const yaml = servicesYaml.stringify();
     const sourceClusterYaml = `source_cluster:\n  endpoint: ${sourceCluster.endpoint}\n  basic_auth:\n    user: ${sourceClusterUser}\n    password: ${sourceClusterPassword}\n`
     expect(yaml).toBe(`${sourceClusterYaml}target_cluster:\n  endpoint: ${targetCluster.endpoint}\n  no_auth: ""\nmetrics_source:\n  cloudwatch:\n`);
-})
+  });
 
-test('Test servicesYaml with rfs backfill can be stringified', () => {
+  test('Test servicesYaml with rfs backfill can be stringified', () => {
     const clusterName = "migration-cluster-name";
     const serviceName = "rfs-service-name";
     const region = "us-east-1"
@@ -62,15 +70,15 @@ test('Test servicesYaml with rfs backfill can be stringified', () => {
     expect(servicesYaml.backfill instanceof RFSBackfillYaml).toBeTruthy();
     const yaml = servicesYaml.stringify();
     expect(yaml).toBe(`metrics_source:\n  cloudwatch:\nbackfill:\n  reindex_from_snapshot:\n    ecs:\n      cluster_name: ${clusterName}\n      service_name: ${serviceName}\n      aws_region: ${region}\n`);
-})
+  });
 
-test('Test servicesYaml without backfill does not include backend section', () => {
+  test('Test servicesYaml without backfill does not include backend section', () => {
     let servicesYaml = new ServicesYaml();
     const yaml = servicesYaml.stringify();
     expect(yaml).toBe(`metrics_source:\n  cloudwatch:\n`);
-})
+  });
 
-test('Test SnapshotYaml for filesystem only includes fs', () => {
+  test('Test SnapshotYaml for filesystem only includes fs', () => {
     let fsSnapshot = new SnapshotYaml();
     fsSnapshot.fs = {"repo_path": "/path/to/shared/volume"}
     const fsSnapshotDict = fsSnapshot.toDict()
@@ -78,9 +86,9 @@ test('Test SnapshotYaml for filesystem only includes fs', () => {
     expect(fsSnapshotDict).toHaveProperty("fs");
     expect(fsSnapshotDict["fs"]).toHaveProperty("repo_path");
     expect(fsSnapshotDict).not.toHaveProperty("s3");
-})
+  });
 
-test('Test SnapshotYaml for s3 only includes s3', () => {
+  test('Test SnapshotYaml for s3 only includes s3', () => {
     let s3Snapshot = new SnapshotYaml();
     s3Snapshot.s3 = {"repo_uri": "s3://repo/path", "aws_region": "us-east-1"}
     const s3SnapshotDict = s3Snapshot.toDict()
@@ -88,16 +96,16 @@ test('Test SnapshotYaml for s3 only includes s3', () => {
     expect(s3SnapshotDict).toHaveProperty("s3");
     expect(s3SnapshotDict["s3"]).toHaveProperty("repo_uri");
     expect(s3SnapshotDict).not.toHaveProperty("fs");
-})
+  });
 
-test('Test that services yaml parameter is created by migration console stack', () => {
+  test('Test that services yaml parameter is created by migration console stack', () => {
     const contextOptions = {
-        vpcEnabled: true,
-        migrationAssistanceEnabled: true,
-        migrationConsoleServiceEnabled: true,
-        sourceClusterEndpoint: "https://test-cluster",
-        reindexFromSnapshotServiceEnabled: true,
-        trafficReplayerServiceEnabled: true
+      vpcEnabled: true,
+      migrationAssistanceEnabled: true,
+      migrationConsoleServiceEnabled: true,
+      sourceClusterEndpoint: "https://test-cluster",
+      reindexFromSnapshotServiceEnabled: true,
+      trafficReplayerServiceEnabled: true
     }
 
     const stacks = createStackComposer(contextOptions)
@@ -107,9 +115,9 @@ test('Test that services yaml parameter is created by migration console stack', 
 
     const valueCapture = new Capture();
     migrationConsoleStackTemplate.hasResourceProperties("AWS::SSM::Parameter", {
-        Type: "String",
-        Name: Match.stringLikeRegexp("/migration/.*/.*/servicesYamlFile"),
-        Value: valueCapture,
+      Type: "String",
+      Name: Match.stringLikeRegexp("/migration/.*/.*/servicesYamlFile"),
+      Value: valueCapture,
     });
     const value = valueCapture.asObject()
     expect(value).toBeDefined();
@@ -127,4 +135,5 @@ test('Test that services yaml parameter is created by migration console stack', 
     const expectedFields = ['source_cluster', 'target_cluster', 'metrics_source', 'backfill', 'snapshot', 'metadata_migration', 'replay', 'kafka'];
     expect(Object.keys(parsedFromYaml).length).toEqual(expectedFields.length)
     expect(new Set(Object.keys(parsedFromYaml))).toEqual(new Set(expectedFields))
-})
+  });
+});
