+++ deployment/cdk/opensearch-service-migration/test/msk-utility-stack.test.ts
@@ -1,15 +1,30 @@
-import {createStackComposer} from "./test-utils";
-import {Match, Template} from "aws-cdk-lib/assertions";
-import {MSKUtilityStack} from "../lib/msk-utility-stack";
+import { createStackComposer } from "./test-utils";
+import { Match, Template } from "aws-cdk-lib/assertions";
+import { MSKUtilityStack } from "../lib/msk-utility-stack";
+import { describe, beforeEach, afterEach, test, expect, jest } from '@jest/globals';
+import { ContainerImage } from "aws-cdk-lib/aws-ecs";
 
-test('Test if mskEnablePublicEndpoints is provided, wait condition and max attempts are set for lambda custom resource', () => {
+jest.mock('aws-cdk-lib/aws-ecr-assets');
+describe('MSKUtilityStack Tests', () => {
+  beforeEach(() => {
+    jest.spyOn(ContainerImage, 'fromDockerImageAsset').mockImplementation(() => ContainerImage.fromRegistry("ServiceImage"));
+  });
+
+  afterEach(() => {
+    jest.clearAllMocks();
+    jest.resetModules();
+    jest.restoreAllMocks();
+    jest.resetAllMocks();
+  });
+
+  test('Test if mskEnablePublicEndpoints is provided, wait condition and max attempts are set for lambda custom resource', () => {
     const contextOptions = {
-        vpcEnabled: true,
-        migrationAssistanceEnabled: true,
-        mskEnablePublicEndpoints: true,
-        mskRestrictPublicAccessTo: "10.0.0.0/32",
-        mskRestrictPublicAccessType: "ipv4",
-        sourceClusterEndpoint: "https://test-cluster",
+      vpcEnabled: true,
+      migrationAssistanceEnabled: true,
+      mskEnablePublicEndpoints: true,
+      mskRestrictPublicAccessTo: "10.0.0.0/32",
+      mskRestrictPublicAccessType: "ipv4",
+      sourceClusterEndpoint: "https://test-cluster",
     }
 
     const stacks = createStackComposer(contextOptions)
@@ -18,23 +33,23 @@ test('Test if mskEnablePublicEndpoints is provided, wait condition and max attem
     const mskUtilityStackTemplate = Template.fromStack(mskUtilityStack)
 
     mskUtilityStackTemplate.hasResource("AWS::Lambda::Function", {
-        Properties: {
-            Environment: {
-                Variables: {
-                    "MAX_ATTEMPTS": "4",
-                    "MSK_ARN": Match.anyValue()
-                }
-            }
+      Properties: {
+        Environment: {
+          Variables: {
+            "MAX_ATTEMPTS": "4",
+            "MSK_ARN": Match.anyValue()
+          }
         }
+      }
     })
     mskUtilityStackTemplate.resourceCountIs("AWS::CloudFormation::WaitConditionHandle", 1)
-})
+  });
 
-test('Test if mskEnablePublicEndpoints is not provided, single run lambda custom resource is created', () => {
+  test('Test if mskEnablePublicEndpoints is not provided, single run lambda custom resource is created', () => {
     const contextOptions = {
-        vpcEnabled: true,
-        migrationAssistanceEnabled: true,
-        sourceClusterEndpoint: "https://test-cluster",
+      vpcEnabled: true,
+      migrationAssistanceEnabled: true,
+      sourceClusterEndpoint: "https://test-cluster",
     }
 
     const stacks = createStackComposer(contextOptions)
@@ -43,13 +58,14 @@ test('Test if mskEnablePublicEndpoints is not provided, single run lambda custom
     const mskUtilityStackTemplate = Template.fromStack(mskUtilityStack)
 
     mskUtilityStackTemplate.hasResource("AWS::Lambda::Function", {
-        Properties: {
-            Environment: {
-                Variables: {
-                    "MSK_ARN": Match.anyValue()
-                }
-            }
+      Properties: {
+        Environment: {
+          Variables: {
+            "MSK_ARN": Match.anyValue()
+          }
         }
+      }
     })
     mskUtilityStackTemplate.resourceCountIs("AWS::CloudFormation::WaitConditionHandle", 0)
-})
+  });
+});
