+++ deployment/cdk/opensearch-service-migration/test/opensearch-domain-stack.test.ts
@@ -1,48 +1,63 @@
-import {Template} from 'aws-cdk-lib/assertions';
-import {OpenSearchDomainStack} from "../lib/opensearch-domain-stack";
-import {NetworkStack} from "../lib/network-stack";
-import {createStackComposer} from "./test-utils";
+import { Template } from 'aws-cdk-lib/assertions';
+import { OpenSearchDomainStack } from "../lib/opensearch-domain-stack";
+import { NetworkStack } from "../lib/network-stack";
+import { createStackComposer } from "./test-utils";
 import { ClusterYaml } from '../lib/migration-services-yaml';
+import { ContainerImage } from "aws-cdk-lib/aws-ecs";
+import { describe, beforeEach, afterEach, test, expect, jest} from '@jest/globals';
 
-test('Test primary context options are mapped with standard data type', () => {
+jest.mock('aws-cdk-lib/aws-ecr-assets');
+describe('OpenSearch Domain Stack Tests', () => {
+  beforeEach(() => {
+    jest.spyOn(ContainerImage, 'fromDockerImageAsset').mockImplementation(() => ContainerImage.fromRegistry("ServiceImage"));
+  });
+
+  afterEach(() => {
+    jest.clearAllMocks();
+    jest.resetModules();
+    jest.restoreAllMocks();
+    jest.resetAllMocks();
+  });
+
+  test('Test primary context options are mapped with standard data type', () => {
     // The cdk.context.json and default-values.json files allow multiple data types
     const contextOptions = {
-        engineVersion: "OS_2.3",
-        domainName: "test-os-domain",
-        dataNodeType: "r6.large.search",
-        dataNodeCount: 5,
-        dedicatedManagerNodeType: "r6g.large.search",
-        dedicatedManagerNodeCount: 3,
-        warmNodeType: "ultrawarm1.medium.search",
-        warmNodeCount: 2,
-        accessPolicies: {
-            "Version": "2012-10-17",
-            "Statement": [{
-                "Effect": "Allow",
-                "Principal": {"AWS": "arn:aws:iam::12345678912:user/test-user"},
-                "Action": "es:ESHttp*",
-                "Resource": "arn:aws:es:us-east-1:12345678912:domain/cdk-os-service-domain/*"
-            }]
-        },
-        fineGrainedManagerUserARN: "arn:aws:iam::12345678912:user/test-user",
-        enforceHTTPS: true,
-        tlsSecurityPolicy: "TLS_1_2",
-        ebsEnabled: true,
-        ebsIops: 4000,
-        ebsVolumeSize: 15,
-        ebsVolumeType: "GP3",
-        encryptionAtRestEnabled: true,
-        encryptionAtRestKmsKeyARN: "arn:aws:kms:us-east-1:12345678912:key/abc123de-4888-4fa7-a508-3811e2d49fc3",
-        loggingAppLogEnabled: true,
-        loggingAppLogGroupARN: "arn:aws:logs:us-east-1:12345678912:log-group:test-log-group:*",
-        nodeToNodeEncryptionEnabled: true,
-        vpcEnabled: true,
-        vpcId: "vpc-123456789abcdefgh",
-        vpcSubnetIds: ["subnet-123456789abcdefgh", "subnet-223456789abcdefgh"],
-        vpcSecurityGroupIds: ["sg-123456789abcdefgh", "sg-223456789abcdefgh"],
-        domainAZCount: 3,
-        domainRemovalPolicy: "DESTROY",
-        sourceClusterEndpoint: "https://test-cluster",
+      engineVersion: "OS_2.3",
+      domainName: "test-os-domain",
+      dataNodeType: "r6.large.search",
+      dataNodeCount: 5,
+      dedicatedManagerNodeType: "r6g.large.search",
+      dedicatedManagerNodeCount: 3,
+      warmNodeType: "ultrawarm1.medium.search",
+      warmNodeCount: 2,
+      accessPolicies: {
+        "Version": "2012-10-17",
+        "Statement": [{
+          "Effect": "Allow",
+          "Principal": { "AWS": "arn:aws:iam::12345678912:user/test-user" },
+          "Action": "es:ESHttp*",
+          "Resource": "arn:aws:es:us-east-1:12345678912:domain/cdk-os-service-domain/*"
+        }]
+      },
+      fineGrainedManagerUserARN: "arn:aws:iam::12345678912:user/test-user",
+      enforceHTTPS: true,
+      tlsSecurityPolicy: "TLS_1_2",
+      ebsEnabled: true,
+      ebsIops: 4000,
+      ebsVolumeSize: 15,
+      ebsVolumeType: "GP3",
+      encryptionAtRestEnabled: true,
+      encryptionAtRestKmsKeyARN: "arn:aws:kms:us-east-1:12345678912:key/abc123de-4888-4fa7-a508-3811e2d49fc3",
+      loggingAppLogEnabled: true,
+      loggingAppLogGroupARN: "arn:aws:logs:us-east-1:12345678912:log-group:test-log-group:*",
+      nodeToNodeEncryptionEnabled: true,
+      vpcEnabled: true,
+      vpcId: "vpc-123456789abcdefgh",
+      vpcSubnetIds: ["subnet-123456789abcdefgh", "subnet-223456789abcdefgh"],
+      vpcSecurityGroupIds: ["sg-123456789abcdefgh", "sg-223456789abcdefgh"],
+      domainAZCount: 3,
+      domainRemovalPolicy: "DESTROY",
+      sourceClusterEndpoint: "https://test-cluster",
     }
 
     const openSearchStacks = createStackComposer(contextOptions)
@@ -52,42 +67,50 @@ test('Test primary context options are mapped with standard data type', () => {
     const networkStack = openSearchStacks.stacks.filter((s) => s instanceof NetworkStack)[0]
     const networkTemplate = Template.fromStack(networkStack)
     assertPrimaryDomainStackTemplate(domainTemplate)
-    // When existing resources are provided the network stack creates no resources
+    // When existing resources are provided, the network stack should only create SSM parameters or osClusterAccessSG
     const resources = networkTemplate.toJSON().Resources;
-    expect(resources === undefined)
-})
+    expect(resources).toBeDefined();
+    for (const resourceKey in resources) {
+      const resourceType = resources[resourceKey].Type;
+      if (resourceType === 'AWS::EC2::SecurityGroup' || resourceType === 'AWS::EC2::SecurityGroupIngress') {
+        expect(resourceKey).toMatch(/^osClusterAccessSG/);
+      } else {
+        expect(resourceType).toBe('AWS::SSM::Parameter');
+      }
+    }
+  });
 
-test('Test primary context options are mapped with only string data type', () => {
+  test('Test primary context options are mapped with only string data type', () => {
     // CDK CLI commands pass all context values as strings
     const contextOptions = {
-        engineVersion: "OS_2.3",
-        domainName: "test-os-domain",
-        dataNodeType: "r6.large.search",
-        dataNodeCount: "5",
-        dedicatedManagerNodeType: "r6g.large.search",
-        dedicatedManagerNodeCount: "3",
-        warmNodeType: "ultrawarm1.medium.search",
-        warmNodeCount: "2",
-        accessPolicies: "{\"Version\":\"2012-10-17\",\"Statement\":[{\"Effect\":\"Allow\",\"Principal\":{\"AWS\":\"arn:aws:iam::12345678912:user/test-user\"},\"Action\":\"es:ESHttp*\",\"Resource\":\"arn:aws:es:us-east-1:12345678912:domain/cdk-os-service-domain/*\"}]}",
-        fineGrainedManagerUserARN: "arn:aws:iam::12345678912:user/test-user",
-        enforceHTTPS: "true",
-        tlsSecurityPolicy: "TLS_1_2",
-        ebsEnabled: "true",
-        ebsIops: "4000",
-        ebsVolumeSize: "15",
-        ebsVolumeType: "GP3",
-        encryptionAtRestEnabled: "true",
-        encryptionAtRestKmsKeyARN: "arn:aws:kms:us-east-1:12345678912:key/abc123de-4888-4fa7-a508-3811e2d49fc3",
-        loggingAppLogEnabled: "true",
-        loggingAppLogGroupARN: "arn:aws:logs:us-east-1:12345678912:log-group:test-log-group:*",
-        nodeToNodeEncryptionEnabled: "true",
-        vpcEnabled: "true",
-        vpcId: "vpc-123456789abcdefgh",
-        vpcSubnetIds: "[\"subnet-123456789abcdefgh\", \"subnet-223456789abcdefgh\"]",
-        vpcSecurityGroupIds: "[\"sg-123456789abcdefgh\", \"sg-223456789abcdefgh\"]",
-        domainAZCount: "3",
-        domainRemovalPolicy: "DESTROY",
-        sourceClusterEndpoint: "https://test-cluster",
+      engineVersion: "OS_2.3",
+      domainName: "test-os-domain",
+      dataNodeType: "r6.large.search",
+      dataNodeCount: "5",
+      dedicatedManagerNodeType: "r6g.large.search",
+      dedicatedManagerNodeCount: "3",
+      warmNodeType: "ultrawarm1.medium.search",
+      warmNodeCount: "2",
+      accessPolicies: "{\"Version\":\"2012-10-17\",\"Statement\":[{\"Effect\":\"Allow\",\"Principal\":{\"AWS\":\"arn:aws:iam::12345678912:user/test-user\"},\"Action\":\"es:ESHttp*\",\"Resource\":\"arn:aws:es:us-east-1:12345678912:domain/cdk-os-service-domain/*\"}]}",
+      fineGrainedManagerUserARN: "arn:aws:iam::12345678912:user/test-user",
+      enforceHTTPS: "true",
+      tlsSecurityPolicy: "TLS_1_2",
+      ebsEnabled: "true",
+      ebsIops: "4000",
+      ebsVolumeSize: "15",
+      ebsVolumeType: "GP3",
+      encryptionAtRestEnabled: "true",
+      encryptionAtRestKmsKeyARN: "arn:aws:kms:us-east-1:12345678912:key/abc123de-4888-4fa7-a508-3811e2d49fc3",
+      loggingAppLogEnabled: "true",
+      loggingAppLogGroupARN: "arn:aws:logs:us-east-1:12345678912:log-group:test-log-group:*",
+      nodeToNodeEncryptionEnabled: "true",
+      vpcEnabled: "true",
+      vpcId: "vpc-123456789abcdefgh",
+      vpcSubnetIds: "[\"subnet-123456789abcdefgh\", \"subnet-223456789abcdefgh\"]",
+      vpcSecurityGroupIds: "[\"sg-123456789abcdefgh\", \"sg-223456789abcdefgh\"]",
+      domainAZCount: "3",
+      domainRemovalPolicy: "DESTROY",
+      sourceClusterEndpoint: "https://test-cluster",
     }
 
     const openSearchStacks = createStackComposer(contextOptions)
@@ -97,22 +120,30 @@ test('Test primary context options are mapped with only string data type', () =>
     const networkStack = openSearchStacks.stacks.filter((s) => s instanceof NetworkStack)[0]
     const networkTemplate = Template.fromStack(networkStack)
     assertPrimaryDomainStackTemplate(domainTemplate)
-    // When existing resources are provided the network stack creates no resources
+    // When existing resources are provided, the network stack should only create SSM parameters or osClusterAccessSG
     const resources = networkTemplate.toJSON().Resources;
-    expect(resources === undefined)
-})
+    expect(resources).toBeDefined();
+    for (const resourceKey in resources) {
+      const resourceType = resources[resourceKey].Type;
+      if (resourceType === 'AWS::EC2::SecurityGroup' || resourceType === 'AWS::EC2::SecurityGroupIngress') {
+        expect(resourceKey).toMatch(/^osClusterAccessSG/);
+      } else {
+        expect(resourceType).toBe('AWS::SSM::Parameter');
+      }
+    }
+  });
 
-test('Test alternate context options are mapped with standard data type', () => {
+  test('Test alternate context options are mapped with standard data type', () => {
     // The cdk.context.json and default-values.json files allow multiple data types
     const contextOptions = {
-        useUnsignedBasicAuth: true,
-        fineGrainedManagerUserName: "admin",
-        fineGrainedManagerUserSecretManagerKeyARN: "arn:aws:secretsmanager:us-east-1:12345678912:secret:master-user-os-pass-123abc",
-        // Fine-grained access requires enforceHTTPS, encryptionAtRest, and nodeToNodeEncryption to be enabled
-        enforceHTTPS: true,
-        encryptionAtRestEnabled: true,
-        nodeToNodeEncryptionEnabled: true,
-        sourceClusterEndpoint: "https://test-cluster",
+      useUnsignedBasicAuth: true,
+      fineGrainedManagerUserName: "admin",
+      fineGrainedManagerUserSecretManagerKeyARN: "arn:aws:secretsmanager:us-east-1:12345678912:secret:master-user-os-pass-123abc",
+      // Fine-grained access requires enforceHTTPS, encryptionAtRest, and nodeToNodeEncryption to be enabled
+      enforceHTTPS: true,
+      encryptionAtRestEnabled: true,
+      nodeToNodeEncryptionEnabled: true,
+      sourceClusterEndpoint: "https://test-cluster",
     }
 
     const openSearchStacks = createStackComposer(contextOptions)
@@ -120,19 +151,19 @@ test('Test alternate context options are mapped with standard data type', () => 
     const domainStack = openSearchStacks.stacks.filter((s) => s instanceof OpenSearchDomainStack)[0]
     const domainTemplate = Template.fromStack(domainStack)
     assertAlternateDomainStackTemplate(domainTemplate)
-})
+  })
 
-test('Test alternate context options are mapped with only string data type', () => {
+  test('Test alternate context options are mapped with only string data type', () => {
     // CDK CLI commands pass all context values as strings
     const contextOptions = {
-        useUnsignedBasicAuth: "true",
-        fineGrainedManagerUserName: "admin",
-        fineGrainedManagerUserSecretManagerKeyARN: "arn:aws:secretsmanager:us-east-1:12345678912:secret:master-user-os-pass-123abc",
-        // Fine-grained access requires enforceHTTPS, encryptionAtRest, and nodeToNodeEncryption to be enabled
-        enforceHTTPS: "true",
-        encryptionAtRestEnabled: "true",
-        nodeToNodeEncryptionEnabled: "true",
-        sourceClusterEndpoint: "https://test-cluster",
+      useUnsignedBasicAuth: "true",
+      fineGrainedManagerUserName: "admin",
+      fineGrainedManagerUserSecretManagerKeyARN: "arn:aws:secretsmanager:us-east-1:12345678912:secret:master-user-os-pass-123abc",
+      // Fine-grained access requires enforceHTTPS, encryptionAtRest, and nodeToNodeEncryption to be enabled
+      enforceHTTPS: "true",
+      encryptionAtRestEnabled: "true",
+      nodeToNodeEncryptionEnabled: "true",
+      sourceClusterEndpoint: "https://test-cluster",
     }
 
     const openSearchStacks = createStackComposer(contextOptions)
@@ -140,11 +171,11 @@ test('Test alternate context options are mapped with only string data type', () 
     const domainStack = openSearchStacks.stacks.filter((s) => s instanceof OpenSearchDomainStack)[0]
     const domainTemplate = Template.fromStack(domainStack)
     assertAlternateDomainStackTemplate(domainTemplate)
-})
+  })
 
-test('Test openAccessPolicy setting creates access policy when enabled', () => {
+  test('Test openAccessPolicy setting creates access policy when enabled', () => {
     const contextOptions = {
-        openAccessPolicyEnabled: true
+      openAccessPolicyEnabled: true
     }
 
     const openSearchStacks = createStackComposer(contextOptions)
@@ -153,12 +184,11 @@ test('Test openAccessPolicy setting creates access policy when enabled', () => {
     const domainTemplate = Template.fromStack(domainStack)
     // Check that openAccessPolicy is created
     domainTemplate.resourceCountIs("Custom::OpenSearchAccessPolicy", 1)
+  })
 
-})
-
-test('Test openAccessPolicy setting does not create access policy when disabled', () => {
+  test('Test openAccessPolicy setting does not create access policy when disabled', () => {
     const contextOptions = {
-        openAccessPolicyEnabled: false
+      openAccessPolicyEnabled: false
     }
 
     const openSearchStacks = createStackComposer(contextOptions)
@@ -167,12 +197,11 @@ test('Test openAccessPolicy setting does not create access policy when disabled'
     const domainTemplate = Template.fromStack(domainStack)
     // Check that openAccessPolicy is not created
     domainTemplate.resourceCountIs("Custom::OpenSearchAccessPolicy", 0)
+  })
 
-})
-
-test('Test openAccessPolicy setting is mapped with string data type', () => {
+  test('Test openAccessPolicy setting is mapped with string data type', () => {
     const contextOptions = {
-        openAccessPolicyEnabled: true
+      openAccessPolicyEnabled: true
     }
 
     const openSearchStacks = createStackComposer(contextOptions)
@@ -181,40 +210,39 @@ test('Test openAccessPolicy setting is mapped with string data type', () => {
     const domainTemplate = Template.fromStack(domainStack)
     // Check that openAccessPolicy is created
     domainTemplate.resourceCountIs("Custom::OpenSearchAccessPolicy", 1)
+  })
 
-})
-
-test( 'Test default stack is created when empty context options are provided for non-required options', () => {
+  test('Test default stack is created when empty context options are provided for non-required options', () => {
     const contextOptions = {
-        dataNodeType: "",
-        dataNodeCount: "",
-        dedicatedManagerNodeType: "",
-        dedicatedManagerNodeCount: "",
-        warmNodeType: "",
-        warmNodeCount: "",
-        accessPolicies: "",
-        useUnsignedBasicAuth: "",
-        fineGrainedManagerUserARN: "",
-        fineGrainedManagerUserName: "",
-        fineGrainedManagerUserSecretManagerKeyARN: "",
-        enforceHTTPS: "",
-        tlsSecurityPolicy: "",
-        ebsEnabled: "",
-        ebsIops: "",
-        ebsVolumeSize: "",
-        ebsVolumeType: "",
-        encryptionAtRestEnabled: "",
-        encryptionAtRestKmsKeyARN: "",
-        loggingAppLogEnabled: "",
-        loggingAppLogGroupARN: "",
-        nodeToNodeEncryptionEnabled: "",
-        vpcEnabled: "",
-        vpcId: "",
-        vpcSubnetIds: "",
-        vpcSecurityGroupIds: "",
-        domainAZCount: "",
-        openAccessPolicyEnabled: "",
-        domainRemovalPolicy: ""
+      dataNodeType: "",
+      dataNodeCount: "",
+      dedicatedManagerNodeType: "",
+      dedicatedManagerNodeCount: "",
+      warmNodeType: "",
+      warmNodeCount: "",
+      accessPolicies: "",
+      useUnsignedBasicAuth: "",
+      fineGrainedManagerUserARN: "",
+      fineGrainedManagerUserName: "",
+      fineGrainedManagerUserSecretManagerKeyARN: "",
+      enforceHTTPS: "",
+      tlsSecurityPolicy: "",
+      ebsEnabled: "",
+      ebsIops: "",
+      ebsVolumeSize: "",
+      ebsVolumeType: "",
+      encryptionAtRestEnabled: "",
+      encryptionAtRestKmsKeyARN: "",
+      loggingAppLogEnabled: "",
+      loggingAppLogGroupARN: "",
+      nodeToNodeEncryptionEnabled: "",
+      vpcEnabled: "",
+      vpcId: "",
+      vpcSubnetIds: "",
+      vpcSecurityGroupIds: "",
+      domainAZCount: "",
+      openAccessPolicyEnabled: "",
+      domainRemovalPolicy: ""
     }
 
     const openSearchStacks = createStackComposer(contextOptions)
@@ -222,9 +250,9 @@ test( 'Test default stack is created when empty context options are provided for
     const domainStack = openSearchStacks.stacks.filter((s) => s instanceof OpenSearchDomainStack)[0]
     const domainTemplate = Template.fromStack(domainStack)
     domainTemplate.resourceCountIs("AWS::OpenSearchService::Domain", 1)
-})
+  })
 
-test('Test targetClusterYaml is populated', () => {
+  test('Test targetClusterYaml is populated', () => {
     const contextOptions = {}
     const openSearchStacks = createStackComposer(contextOptions)
 
@@ -233,6 +261,7 @@ test('Test targetClusterYaml is populated', () => {
     expect(yaml).toBeDefined()
     expect(yaml).toBeInstanceOf(ClusterYaml)
     expect(yaml.endpoint).toBeDefined()
+  })
 })
 
 /*
@@ -240,65 +269,65 @@ test('Test targetClusterYaml is populated', () => {
  * which should not interfere with resource properties of other settings in the set
  */
 function assertPrimaryDomainStackTemplate(template: Template) {
-    // Check that accessPolicies policy is created
-    template.resourceCountIs("Custom::OpenSearchAccessPolicy", 1)
-    template.resourceCountIs("AWS::OpenSearchService::Domain", 1)
-    template.hasResourceProperties("AWS::OpenSearchService::Domain", {
-        EngineVersion: "OpenSearch_2.3",
-        DomainName: "test-os-domain",
-        AdvancedSecurityOptions: {
-            Enabled: true,
-            MasterUserOptions: {
-                MasterUserARN: "arn:aws:iam::12345678912:user/test-user"
-            }
-        },
-        ClusterConfig: {
-            DedicatedMasterCount: 3,
-            DedicatedMasterEnabled: true,
-            DedicatedMasterType: "r6g.large.search",
-            InstanceCount: 5,
-            InstanceType: "r6.large.search",
-            WarmCount: 2,
-            WarmType: "ultrawarm1.medium.search",
-            ZoneAwarenessConfig: {
-                AvailabilityZoneCount: 3
-            },
-            ZoneAwarenessEnabled: true
-        },
-        DomainEndpointOptions: {
-            EnforceHTTPS: true,
-            TLSSecurityPolicy: "Policy-Min-TLS-1-2-2019-07"
-        },
-        EBSOptions: {
-            EBSEnabled: true,
-            Iops: 4000,
-            VolumeSize: 15,
-            VolumeType: "gp3"
-        },
-        EncryptionAtRestOptions: {
-            Enabled: true,
-            KmsKeyId: "abc123de-4888-4fa7-a508-3811e2d49fc3"
-        },
-        LogPublishingOptions: {
-            ES_APPLICATION_LOGS: {
-                CloudWatchLogsLogGroupArn: "arn:aws:logs:us-east-1:12345678912:log-group:test-log-group:*",
-                Enabled: true
-            }
-        },
-        /*
-        * Only checking that the VPCOptions object is added here as normally the provided vpcId will perform a lookup to
-        * determine these options, but seems to be auto mocked here
-        */
-        VPCOptions: {},
-        NodeToNodeEncryptionOptions: {
-            Enabled: true
-        }
-    })
-    // Check our removal policy has been added
-    template.hasResource("AWS::OpenSearchService::Domain", {
-        DeletionPolicy: "Delete",
-        UpdateReplacePolicy: "Delete"
-    })
+  // Check that accessPolicies policy is created
+  template.resourceCountIs("Custom::OpenSearchAccessPolicy", 1)
+  template.resourceCountIs("AWS::OpenSearchService::Domain", 1)
+  template.hasResourceProperties("AWS::OpenSearchService::Domain", {
+    EngineVersion: "OpenSearch_2.3",
+    DomainName: "test-os-domain",
+    AdvancedSecurityOptions: {
+      Enabled: true,
+      MasterUserOptions: {
+        MasterUserARN: "arn:aws:iam::12345678912:user/test-user"
+      }
+    },
+    ClusterConfig: {
+      DedicatedMasterCount: 3,
+      DedicatedMasterEnabled: true,
+      DedicatedMasterType: "r6g.large.search",
+      InstanceCount: 5,
+      InstanceType: "r6.large.search",
+      WarmCount: 2,
+      WarmType: "ultrawarm1.medium.search",
+      ZoneAwarenessConfig: {
+        AvailabilityZoneCount: 3
+      },
+      ZoneAwarenessEnabled: true
+    },
+    DomainEndpointOptions: {
+      EnforceHTTPS: true,
+      TLSSecurityPolicy: "Policy-Min-TLS-1-2-2019-07"
+    },
+    EBSOptions: {
+      EBSEnabled: true,
+      Iops: 4000,
+      VolumeSize: 15,
+      VolumeType: "gp3"
+    },
+    EncryptionAtRestOptions: {
+      Enabled: true,
+      KmsKeyId: "abc123de-4888-4fa7-a508-3811e2d49fc3"
+    },
+    LogPublishingOptions: {
+      ES_APPLICATION_LOGS: {
+        CloudWatchLogsLogGroupArn: "arn:aws:logs:us-east-1:12345678912:log-group:test-log-group:*",
+        Enabled: true
+      }
+    },
+    /*
+    * Only checking that the VPCOptions object is added here as normally the provided vpcId will perform a lookup to
+    * determine these options, but seems to be auto mocked here
+    */
+    VPCOptions: {},
+    NodeToNodeEncryptionOptions: {
+      Enabled: true
+    }
+  })
+  // Check our removal policy has been added
+  template.hasResource("AWS::OpenSearchService::Domain", {
+    DeletionPolicy: "Delete",
+    UpdateReplacePolicy: "Delete"
+  })
 }
 
 /*
